A-10 DSMS: use GetNumClicksForWraparoundSetting
