+++ JAFDTC/Models/A10C/DSMS/MunitionSettings.cs
@@ -24,6 +24,71 @@ using System.Text.Json.Serialization;
 
 namespace JAFDTC.Models.A10C.DSMS
 {
+    // defines the possible delivery mode settings
+    // must match combobox order in ui
+    public enum DeliveryModes
+    {
+        CCIP = 0,
+        CCRP = 1
+    }
+
+    // defines the possible escape maneuver settings
+    // must match combobox order in ui
+    public enum EscapeManeuvers
+    {
+        NONE = 0,
+        CLB = 1, // jet default
+        TRN = 2,
+        TLT = 3
+    }
+
+    // defines the possible release mode settings
+    // must match combobox order in ui
+    public enum ReleaseModes
+    {
+        SGL = 0, // jet default
+        PRS = 1,
+        RIP_SGL = 2,
+        RIP_PRS = 3
+    }
+
+    // defines the possible height of function (HOF) settings
+    // must match combobox order in ui
+    public enum HOFOptions
+    {
+        HOF_300 = 0,
+        HOF_500 = 1,
+        HOF_700 = 2,
+        HOF_900 = 3,
+        HOF_1200 = 4,
+        HOF_1500 = 5,
+        HOF_1800 = 6, // jet default
+        HOF_2200 = 7,
+        HOF_2600 = 8,
+        HOF_3000 = 9
+    }
+
+    // defines the RPM settings
+    // must match combobox order in ui
+    public enum RPMOptions
+    {
+        RPM_0 = 0,
+        RPM_500 = 1,
+        RPM_1000 = 2,
+        RPM_1500 = 3, // jet default
+        RPM_2000 = 4,
+        RPM_2500 = 5
+    }
+
+    // defines the Fuze settings
+    // must match combobox order in ui
+    public enum FuzeOptions
+    {
+        NoseTail = 0, // jet default
+        Nose = 1,
+        Tail = 2
+    }
+
     public class MunitionSettings : BindableObject, ISystem
     {
         private string _autoLase;                              // string (boolean)
